+++ packages/client/hmi-client/src/components/workflow/ops/model-comparison/mod.ts
@@ -1,6 +1,6 @@
 import { ModelComparisonOperation as operation } from './model-comparison-operation';
 import node from './tera-model-comparison-node.vue';
-import drilldown from './tera-model-comparison.vue';
+import drilldown from './tera-model-comparison-drilldown.vue';
 
 const name = operation.name;
